remove .py extensions from entry points
Executable scripts get installed in a separate `bin/` folder. If executable scripts end with the `.py` extension, they will clash with the `.py` source files in the `site-packags/` folder. This means that `import` Python directive will load the executable scripts instead of the source files in `site-packages/`.

This is a problem because some distributions (like NixOS) wrap these executable scripts into shell scripts. Now when Python tries to import these scripts it will fail (because the code in those files is now a shell script rather than Python).